@@ -1,10 +1,14 @@
 package com.devsuperior.uri2621;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import com.devsuperior.uri2621.dto.ProductDto;
+import com.devsuperior.uri2621.projections.ProductMinProjection;
 import com.devsuperior.uri2621.repositories.ProductRepository;
 
 @SpringBootApplication
@@ -12,13 +16,19 @@ public class Uri2621Application implements CommandLineRunner {
 
 	@Autowired
 	private ProductRepository repository;
-	
+
 	public static void main(String[] args) {
 		SpringApplication.run(Uri2621Application.class, args);
 	}
 
 	@Override
 	public void run(String... args) throws Exception {
-		
+
+		List<ProductMinProjection> list = repository.search1(10, 20, "P");
+		list.stream().forEach(x -> System.out.println(x.getName()));
+
+		List<ProductDto> list2 = repository.search2(10, 20, "P");
+		list2.stream().forEach(x -> System.out.println(x));
+
 	}
 }
